Refactored AbstractBox2dObject
